refactor: primary 버튼 색상 공용 클래스 "primary-btn" 생성
- active, hover경우에만
