+++ src/components/myScrap/Button.tsx
@@ -6,7 +6,7 @@ export default function Button({ text, onClick, type = 'button' }: ButtonProps) 
       id={`${text === '콘텐츠 모으러 가기' ? 'btn-content-collect' : 'btn-term-collect'}`}
       type={type}
       onClick={onClick}
-      className='h-12 w-full max-w-[343px] rounded-[10px] bg-custom-gray-dark font-medium text-primary transition hover:bg-hover-80 hover:text-green-hover md:text-sm'
+      className='primary-btn h-12 w-full max-w-[343px] rounded-[10px]'
     >
       {text}
     </button>
